Cleaning Deleted TorchFunctions and moved everything in Functions and DefaultActivity
